Backend/project/requirements.txt
asgiref==3.7.2
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-bulk==0.2.1
drf-writable-nested==0.7.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
